[wip] add tests for refund
