+++ src/worker/jobs/find-claim-swap-tx.js
@@ -1,19 +1,23 @@
 const Order = require('../../models/Order')
 const debug = require('debug')('liquality:agent:worker')
 
+const TEST_ENV = process.env.NODE_ENV === 'test'
+
 async function findClaim (order, lastScannedBlock, currentBlock) {
   const newBlocksExist = !lastScannedBlock || (currentBlock > lastScannedBlock)
   if (newBlocksExist) {
     let blockNumber = lastScannedBlock ? lastScannedBlock + 1 : currentBlock
+
     for (;blockNumber <= currentBlock; blockNumber++) {
       const claimTx = await order.toClient().swap.findClaimSwapTransaction(
         order.toFundHash,
         order.toAddress,
         order.toCounterPartyAddress,
         order.secretHash,
-        order.nodeExpiration,
+        order.nodeSwapExpiration,
         blockNumber
       )
+
       if (claimTx) return claimTx
     }
   }
@@ -25,21 +29,20 @@ module.exports = agenda => async job => {
   const order = await Order.findOne({ orderId: data.orderId }).exec()
   if (!order) return
 
-  const currentBlock = await order.toClient().chain.getBlockHeight() // TODO: order (initator) should provide start block
-
+  const currentBlock = await order.toClient().chain.getBlockHeight()
   const claimTx = await findClaim(order, data.lastScannedBlock, currentBlock)
-  const lastScannedBlock = currentBlock // TODO: persist last scanned block to prevent situation where agent going offline loses state.
 
   if (!claimTx) {
     const block = await order.toClient().chain.getBlockByNumber(currentBlock)
-    if (block.timestamp <= order.nodeExpiration) {
-      // TODO: use block times as schedule?
-      await agenda.schedule('in 10 seconds', 'find-claim-swap-tx', { orderId: data.orderId, lastScannedBlock })
-      return
-    } else {
+
+    if (block.timestamp >= order.nodeSwapExpiration) {
       await agenda.now('agent-refund', { orderId: order.orderId })
-      return
+    } else {
+      const when = TEST_ENV ? 'in 2 seconds' : 'in 10 seconds'
+      await agenda.schedule(when, 'find-claim-swap-tx', { orderId: data.orderId, lastScannedBlock: currentBlock })
     }
+
+    return
   }
 
   order.secret = claimTx.secret
