+++ src/worker/jobs/reciprocate-init-swap.js
@@ -7,7 +7,17 @@ module.exports = agenda => async job => {
   const order = await Order.findOne({ orderId: data.orderId }).exec()
   if (!order) return
 
-  const tx = await order.toClient().swap.initiateSwap(order.toAmount, order.toAddress, order.toCounterPartyAddress, order.secretHash, order.nodeExpiration)
+  const currentBlock = await order.fromClient().chain.getBlockHeight()
+  const block = await order.fromClient().chain.getBlockByNumber(currentBlock)
+
+  if (block.timestamp >= order.swapExpiration) { // no need to continue
+    debug(`Order ${order.orderId} expired due to swapExpiration`)
+    order.status = 'SWAP_EXPIRED'
+    await order.save()
+    return
+  }
+
+  const tx = await order.toClient().swap.initiateSwap(order.toAmount, order.toAddress, order.toCounterPartyAddress, order.secretHash, order.nodeSwapExpiration)
   debug('Initiated funding transaction', order.orderId, tx)
 
   order.toFundHash = tx
