+++ test/docker/miner.Dockerfile
@@ -1,4 +1,9 @@
 FROM alpine
-COPY mine.sh /home/mine.sh
+
 WORKDIR /home
+
+RUN apk add curl curl-dev
+
+COPY mine.sh /home/mine.sh
+
 CMD /bin/sh mine.sh
