+++ test/docker/proxy.Dockerfile
@@ -1,4 +1,9 @@
 FROM node:8
+
 WORKDIR /home/node/app
+
+RUN npm install http-proxy
+
 COPY corsproxy.js /home/node/app/corsproxy.js
-CMD npm install http-proxy && node corsproxy.js
+
+CMD node corsproxy.js
